cmd/.../scorecard: move scorecard libraries out of cmd
